🆕 feat(repo_service_grpc.pb.go): add generated code for gRPC service definition
The commit adds the generated code for the `RepoService` gRPC service definition. This code was generated using `protoc-gen-go-grpc` version 1.2.0 and `protoc` version 3.12.4. The generated code includes the client and server interfaces for the `RepoService` service, as well as the implementation of these interfaces.

This commit is necessary to provide the necessary infrastructure for implementing and using the `RepoService` gRPC service.